декомпозиция стилей и интерфейсов
