+++ src/models/IDoc.ts
@@ -1,0 +1,4 @@
+export interface IDoc {
+    name: string;
+    picture: string;
+}
