+++ src/models/IMyForm.ts
@@ -1,0 +1,4 @@
+export interface IMyForm {
+    name: string;
+    picture: string;
+}
