+++ src/models/IUniversity.ts
@@ -1,0 +1,4 @@
+export interface IUniversity {
+    country: string;
+    name: string;
+}
